backend/pyproject.toml
[tool.poetry]
name = "databae"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
uagents = "^0.16.2"
fastapi = "^0.115.2"
uvicorn = "^0.30.1"
mysql-connector-python = "^9.1.0"
groq = "^0.11.0"
pydantic = "2.8.2"
instructor = "^1.6.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
